Add auto unitary tests
