+++ package.json
@@ -3,11 +3,11 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "ts-node-dev --respawn --transpile-only api/index.ts",
-    "build": "tsc",
-    "start": "ts-node api/index.ts"
-  },
+  "test": "jest --passWithNoTests",
+  "dev": "ts-node-dev --respawn --transpile-only api/index.ts",
+  "build": "tsc",
+  "start": "ts-node api/index.ts"
+},
   "keywords": [],
   "author": "",
   "license": "ISC",
