+++ src/routes/UsersRoutes.ts
@@ -1,24 +1,10 @@
-import { Router } from "express";
-import {
-  getAllUsers,
-  getUserById,
-  createUser,
-  updateUser,
-  deleteUser,
-  restoreUser,
-} from "../controllers/user.controller";
+import app from '../../api/app';
+import { connectDB } from '../config/db';
+import router from './AuthRoutes';
 
-import { authenticateToken } from "../middlewares/authmiddlewares";
+connectDB();
 
-const router: Router = Router();
-
-router.get("/", getAllUsers); // rota pública
-router.get("/:id", getUserById); // rota pública
-router.post("/register", createUser); // rota pública 
-
-// rotas privadas
-router.put("/:id", authenticateToken, updateUser);
-router.delete("/:id", authenticateToken, deleteUser);
-router.patch("/:id/restore", authenticateToken, restoreUser);
-
+app.listen(process.env.PORT, () => {
+  console.log(`Servidor rodando na porta ${process.env.PORT}`);
+});
 export default router;
